Improve install candidate selection     Florian Boor <florian.boor@kernelconcepts.de> Hello,
I have noticed that opkg does not resolve package selection properly if
the same
package is available in more than one compatible architecture.
Failing having installable candidates available is not necessary in this
case:
Either the packages are from feeds with different priorities - in this
case we
can select by priority - or the priorities are the same and it does not
matter
which one gets installed.
Another problem is that in some situations the package lists seem to
contain
duplicates for no obvious reason. (e.g. one package available from two
feeds
makes four entries in the package lists.)
The patch implements a workaround for this by filtering out duplicates
when the
candidate lists are created and makes opkg select an install candidate
per feed
priority.

I have tested the changes with OpenEmbedded and it does not seem to
cause any
obvious problems.
Could I get some opinions about the solution? :)

Greetings

Florian
-----------
Thanks to Florian

git-svn-id: http://opkg.googlecode.com/svn/trunk@219 e8e0d7a0-c8d9-11dd-a880-a1081c7ac358