Add the --select-higher-version option
If there were more than one candidate which had the same pkg name in the
candidate list, for example, the same pkg with different versions, then
it would use the last one which was the highest version one in the past,
but it will use the higher arch priority one now.

Add the "--select-higher-version" option to let it use the higher
version package when enabled. the default is no.

Signed-off-by: Martin Jansa <Martin.Jansa@gmail.com>